fix ping failed metrics
